+++ packages/orq-rc/src/orq_ai_sdk/models/retrieveidentityop.py
@@ -103,7 +103,7 @@ class RetrieveIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-29T10:45:07.477Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-29T13:22:33.026Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")